Hu03 029 mg

+++ src/main/java/com/scoutmanagement/service/implementation/ActividadService.java
@@ -4,14 +4,17 @@ import com.scoutmanagement.dto.ActividadDTO;
 import com.scoutmanagement.persistence.model.Actividad;
 import com.scoutmanagement.persistence.model.Rama;
 import com.scoutmanagement.persistence.repository.ActividadRepository;
+import com.scoutmanagement.persistence.repository.AsistenciaRepository;
 import com.scoutmanagement.service.interfaces.IActividadService;
 import com.scoutmanagement.util.exception.ServiceException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
+import java.time.temporal.TemporalAdjusters;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -20,6 +23,9 @@ public class ActividadService implements IActividadService {
 
     @Autowired
     private ActividadRepository actividadRepository;
+
+    @Autowired
+    private AsistenciaRepository asistenciaRepository;
 
     @Override
     public Optional<Actividad> findById(Long id) {
@@ -49,7 +55,12 @@ public class ActividadService implements IActividadService {
     }
 
     @Override
+    @Transactional
     public void eliminarActividad(Long id) {
+        if (!actividadRepository.existsById(id)) {
+            throw new IllegalArgumentException("Actividad no encontrada con ID: " + id);
+        }
+        asistenciaRepository.deleteByActividadId(id);
         actividadRepository.deleteById(id);
     }
 
@@ -117,8 +128,14 @@ public class ActividadService implements IActividadService {
     }
 
     public Long contarActividadesEstaSemana() {
-        LocalDate finSemana = LocalDate.now().with(DayOfWeek.SUNDAY);
-        return actividadRepository.contarActividadesEstaSemana(finSemana);
+        LocalDate inicioSemana = LocalDate.now().with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY));
+        LocalDate finSemana = LocalDate.now().with(TemporalAdjusters.nextOrSame(DayOfWeek.SUNDAY));
+        return actividadRepository.contarActividadesEstaSemana(inicioSemana,finSemana);
+    }
+
+    @Override
+    public List<Actividad> obtenerTresProximasActividades() {
+        return actividadRepository.findTop3ByFechaGreaterThanEqualOrderByFechaAsc(LocalDate.now());
     }
 
 }
